requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
pydantic>=1.8.2
pydantic-settings>=2.0.0
langchain>=0.0.300
pandas>=1.3.3
numpy>=1.21.2
plotly>=5.3.1
python-dotenv>=0.19.0
pytest>=6.2.5
httpx>=0.18.2
openai>=0.27.0
git+https://github.com/modelcontextprotocol/mcp-python.git#egg=mcp-python
